refactor(auth): move openidcredentials to delegation sub-module
