+++ src/libs/auth/src/delegation/get.rs
@@ -4,8 +4,8 @@ use crate::delegation::types::{
 use crate::delegation::utils::seed::calculate_seed;
 use crate::delegation::utils::signature::{build_signature_inputs, build_signature_msg};
 use crate::delegation::utils::targets::build_targets;
+use crate::openid::delegation::types::interface::{OpenIdCredential, OpenIdCredentialKey};
 use crate::openid::delegation::types::provider::OpenIdDelegationProvider;
-use crate::openid::types::interface::{OpenIdCredential, OpenIdCredentialKey};
 use crate::state::get_salt;
 use crate::state::services::read_state;
 use crate::strategies::{AuthCertificateStrategy, AuthHeapStrategy};
