+++ src/libs/auth/src/openid/delegation/verify.rs
@@ -1,9 +1,9 @@
+use crate::openid::delegation::types::errors::VerifyOpenidCredentialsError;
+use crate::openid::delegation::types::interface::OpenIdCredential;
 use crate::openid::delegation::types::provider::OpenIdDelegationProvider;
 use crate::openid::jwkset::{get_jwks, get_or_refresh_jwks};
 use crate::openid::jwt::types::cert::Jwks;
 use crate::openid::jwt::{unsafe_find_jwt_provider, verify_openid_jwt};
-use crate::openid::types::errors::VerifyOpenidCredentialsError;
-use crate::openid::types::interface::OpenIdCredential;
 use crate::openid::types::provider::OpenIdProvider;
 use crate::openid::utils::build_nonce;
 use crate::state::types::config::{OpenIdProviderClientId, OpenIdProviders};
