+++ src/libs/auth/src/openid/impls.rs
@@ -1,37 +1,9 @@
 use crate::openid::jwt::types::cert::Jwks;
-use crate::openid::jwt::types::token::Claims;
-use crate::openid::types::interface::{OpenIdCredential, OpenIdCredentialKey};
 use crate::openid::types::provider::{OpenIdCertificate, OpenIdProvider};
 use ic_cdk::api::time;
-use jsonwebtoken::TokenData;
 use junobuild_shared::data::version::next_version;
 use junobuild_shared::types::state::{Version, Versioned};
 use std::fmt::{Display, Formatter, Result as FmtResult};
-
-impl From<TokenData<Claims>> for OpenIdCredential {
-    fn from(token: TokenData<Claims>) -> Self {
-        Self {
-            sub: token.claims.sub,
-            iss: token.claims.iss,
-            email: token.claims.email,
-            name: token.claims.name,
-            given_name: token.claims.given_name,
-            family_name: token.claims.family_name,
-            preferred_username: token.claims.preferred_username,
-            picture: token.claims.picture,
-            locale: token.claims.locale,
-        }
-    }
-}
-
-impl<'a> From<&'a OpenIdCredential> for OpenIdCredentialKey<'a> {
-    fn from(credential: &'a OpenIdCredential) -> Self {
-        Self {
-            sub: &credential.sub,
-            iss: &credential.iss,
-        }
-    }
-}
 
 impl OpenIdProvider {
     pub fn jwks_url(&self) -> &'static str {
