+++ src/libs/satellite/src/auth/delegation.rs
@@ -4,8 +4,8 @@ use junobuild_auth::delegation::types::{
     GetDelegationError, GetDelegationResult, OpenIdGetDelegationArgs, OpenIdPrepareDelegationArgs,
     PrepareDelegationError, PreparedDelegation,
 };
+use junobuild_auth::openid::delegation::types::interface::OpenIdCredential;
 use junobuild_auth::openid::delegation::types::provider::OpenIdDelegationProvider;
-use junobuild_auth::openid::types::interface::OpenIdCredential;
 use junobuild_auth::state::types::config::OpenIdProviders;
 use junobuild_auth::{delegation, openid};
 
@@ -22,14 +22,15 @@ pub async fn openid_prepare_delegation(
     args: &OpenIdPrepareDelegationArgs,
     providers: &OpenIdProviders,
 ) -> OpenIdPrepareDelegationResult {
-    let (credential, provider) = match openid::delegation::verify_openid_credentials_with_jwks_renewal(
-        &args.jwt, &args.salt, providers, &AuthHeap,
-    )
-    .await
-    {
-        Ok(value) => value,
-        Err(err) => return Err(PrepareDelegationError::from(err)),
-    };
+    let (credential, provider) =
+        match openid::delegation::verify_openid_credentials_with_jwks_renewal(
+            &args.jwt, &args.salt, providers, &AuthHeap,
+        )
+        .await
+        {
+            Ok(value) => value,
+            Err(err) => return Err(PrepareDelegationError::from(err)),
+        };
 
     let result = delegation::openid_prepare_delegation(
         &args.session_key,
@@ -46,12 +47,13 @@ pub fn openid_get_delegation(
     args: &OpenIdGetDelegationArgs,
     providers: &OpenIdProviders,
 ) -> GetDelegationResult {
-    let (credential, provider) = match openid::delegation::verify_openid_credentials_with_cached_jwks(
-        &args.jwt, &args.salt, providers, &AuthHeap,
-    ) {
-        Ok(value) => value,
-        Err(err) => return Err(GetDelegationError::from(err)),
-    };
+    let (credential, provider) =
+        match openid::delegation::verify_openid_credentials_with_cached_jwks(
+            &args.jwt, &args.salt, providers, &AuthHeap,
+        ) {
+            Ok(value) => value,
+            Err(err) => return Err(GetDelegationError::from(err)),
+        };
 
     delegation::openid_get_delegation(
         &args.session_key,
